Reorganize core orders for Python
